requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==5.0.3
djangorestframework==3.15.0
grpcio==1.62.1
lnurl==0.4.2
python-dotenv==1.0.1
Requests==2.31.0
secp256k1==0.14.0
segno==1.6.1
